test(combine-api): split tests into separate files
